// FordTube.Web/src/app/core/xapi/enums/definition-type.enum.ts
export const enum DefinitionType {
  ALERT = 'http://activitystrea.ms/schema/1.0/alert',
  APPLICATION = 'http://activitystrea.ms/schema/1.0/application',
  ARTICLE = 'http://activitystrea.ms/schema/1.0/article',
  AUDIO = 'http://activitystrea.ms/schema/1.0/audio',
  BADGE = 'http://activitystrea.ms/schema/1.0/badge',
  BINARY = 'http://activitystrea.ms/schema/1.0/binary',
  BOOKMARK = 'http://activitystrea.ms/schema/1.0/bookmark',
  COLLECTION = 'http://activitystrea.ms/schema/1.0/collection',
  CATEGORY = 'http://id.tincanapi.com/activitytype/category',
  COMMENT = 'http://activitystrea.ms/schema/1.0/comment',
  DEVICE = 'http://activitystrea.ms/schema/1.0/device',
  EVENT = 'http://activitystrea.ms/schema/1.0/event',
  FILE = 'http://activitystrea.ms/schema/1.0/file',
  GAME = 'http://activitystrea.ms/schema/1.0/game',
  GROUP = 'http://activitystrea.ms/schema/1.0/group',
  IMAGE = 'http://activitystrea.ms/schema/1.0/image',
  ISSUE = 'http://activitystrea.ms/schema/1.0/issue',
  JOB = 'http://activitystrea.ms/schema/1.0/job',
  NOTE = 'http://activitystrea.ms/schema/1.0/note',
  OFFER = 'http://activitystrea.ms/schema/1.0/offer',
  ORGANIZATION = 'http://activitystrea.ms/schema/1.0/organization',
  PAGE = 'http://activitystrea.ms/schema/1.0/page',
  PERMISSION = 'http://activitystrea.ms/schema/1.0/permission',
  PERSON = 'http://activitystrea.ms/schema/1.0/person',
  PLACE = 'http://activitystrea.ms/schema/1.0/place',
  PROCESS = 'http://activitystrea.ms/schema/1.0/process',
  PRODUCT = 'http://activitystrea.ms/schema/1.0/product',
  QUESTION = 'http://activitystrea.ms/schema/1.0/question',
  REVIEW = 'http://activitystrea.ms/schema/1.0/review',
  ROLE = 'http://activitystrea.ms/schema/1.0/role',
  SERVICE = 'http://activitystrea.ms/schema/1.0/service',
  TASK = 'http://activitystrea.ms/schema/1.0/task',
  TEAM = 'http://activitystrea.ms/schema/1.0/team',
  VIDEO = 'http://activitystrea.ms/schema/1.0/video'
}